Format ioctl uses key, not value
The format ioctl kernel code should not access userspace pointer directly
when accessing file name, it must copy to kernel space first.

The format command uses key for the filename, not value, because the
filename is a zero-terminated string, just like other keys. This
simplifies error checking in kernel ioctl code, where all commands use
the the same copy-key-string-to-kernel code.